+++ crates/miden-tx/src/executor/mod.rs
@@ -39,6 +39,7 @@ pub use notes_checker::{
     MAX_NUM_CHECKER_NOTES,
     NoteConsumptionChecker,
     NoteConsumptionInfo,
+    SuccessfulNote,
 };
 
 mod program_executor;