+++ crates/miden-tx/src/lib.rs
@@ -16,6 +16,7 @@ pub use executor::{
     NoteConsumptionChecker,
     NoteConsumptionInfo,
     ProgramExecutor,
+    SuccessfulNote,
     TransactionExecutor,
     TransactionExecutorHost,
 };